requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
nltk==3.6.3
textblob==0.15.3
python-dotenv==0.19.0
requests==2.26.0 